requirements.txt
pandas
numpy
requests
ta
python-telegram-bot
